feat(evals): `eval_deepagent` dispatcher and shared mock tool registry
Extract all mock tool definitions from the eval test suite into a shared
`deepagents_evals.mock_tools` package and introduce an `EvalSpec`-based
registry (`deepagents_evals.eval_registry`) that maps every eval test name
to its exact `create_deep_agent` configuration.

Register a new `eval_deepagent` graph in `langgraph.json` whose factory,
`make_eval_graph`, reads `configurable["eval_name"]`, looks up the
corresponding `EvalSpec`, and calls `spec.build(model)`. This gives Harbor
a single entry point that can run any eval without per-eval graph
factories.

The pytest test files now import tools from `deepagents_evals.mock_tools`
instead of defining them inline, eliminating ~1,270 lines of duplicated
code. Both pytest and Harbor consume the same builder, so the
`create_deep_agent` call for a given eval exists in exactly one place.
This is the foundational step toward making pytest a thin wrapper around
Harbor.

The registry covers all 109 evals across 11 test modules:
file_operations, memory, memory_multiturn, skills, todos, system_prompt,
followup_quality, subagents, tool_selection, tool_usage_relational, and
tool_usage_incident_graph.
